stylish_backend/stylish/requirements.txt
aioredis==2.0.1
aioredis3==2.0.1
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
Brotli==1.1.0
CacheControl==0.14.2
cachetools==5.5.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cmdstanpy==1.2.5
contourpy==1.3.2
cron-descriptor==1.4.5
cryptography==44.0.1
cssselect2==0.7.0
cycler==0.12.1
distro==1.9.0
Django==5.1.6
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-filter==25.1
django-redis==5.4.0
django-rest-framework==0.1.0
django-storages==1.14.4
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
Faker==37.4.0
firebase-admin==6.6.0
fonttools==4.56.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0rc1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.16.0
holidays==0.75
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
jiter==0.10.0
joblib==1.4.2
kiwisolver==1.4.8
kombu==5.4.2
matplotlib==3.10.3
msgpack==1.1.0
numpy==2.2.3
openai==1.91.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
prophet==1.1.7
proto-plus==1.26.0
protobuf==5.29.3
psycopg-binary==3.2.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==1.10.22
pydyf==0.11.0
PyJWT==2.10.1
pyparsing==3.2.1
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stanio==0.5.1
tenacity==9.1.2
threadpoolctl==3.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==64.0
webencodings==0.5.1
zopfli==0.2.3.post1
